Few refactor afer PR review
