+++ frontend/src/features/CustomZone/components/CustomZones/index.tsx
@@ -3,18 +3,18 @@ import { useCallback, useEffect, useMemo, useState } from 'react'
 import styled from 'styled-components'
 
 import { CustomZone } from './CustomZone'
-import { COLORS } from '../../../constants/constants'
-import { LayerType } from '../../../domain/entities/layers/constants'
-import LayerSlice from '../../../domain/shared_slices/Layer'
-import { useMainAppDispatch } from '../../../hooks/useMainAppDispatch'
-import { useMainAppSelector } from '../../../hooks/useMainAppSelector'
-import { ChevronIcon } from '../../commonStyles/icons/ChevronIcon.style'
-import { initDragAndDrop } from '../useCases/initDragAndDrop'
-import { initLayer } from '../useCases/initLayer'
-import { remove } from '../useCases/remove'
-import { showOrHide } from '../useCases/showOrHide'
+import { COLORS } from '../../../../constants/constants'
+import { LayerType } from '../../../../domain/entities/layers/constants'
+import LayerSlice from '../../../../domain/shared_slices/Layer'
+import { useMainAppDispatch } from '../../../../hooks/useMainAppDispatch'
+import { useMainAppSelector } from '../../../../hooks/useMainAppSelector'
+import { ChevronIcon } from '../../../commonStyles/icons/ChevronIcon.style'
+import { initDragAndDrop } from '../../useCases/initDragAndDrop'
+import { initLayer } from '../../useCases/initLayer'
+import { remove } from '../../useCases/remove'
+import { showOrHide } from '../../useCases/showOrHide'
 
-import type { LayerSliceNamespace } from '../../../domain/entities/layers/types'
+import type { LayerSliceNamespace } from '../../../../domain/entities/layers/types'
 
 export type CustomZonesProps = {
   hideLayersListWhenSearching?: boolean
